+++ src/db/mod.rs
@@ -1,9 +1,5 @@
-mod manage_db;
-mod sub_func;
-mod tx_manager;
-mod verifier;
+mod setup;
+mod update;
 
-pub use manage_db::{add_new_tx_methods, add_tags_column, create_db};
-pub use sub_func::*;
-pub use tx_manager::*;
-pub use verifier::*;
+pub use setup::*;
+pub use update::*;
