+++ src/utility/traits/mod.rs
@@ -1,0 +1,3 @@
+mod verifier;
+
+pub use verifier::StatusChecker;
